ui refactor
